Changes to make Heightfield data more general purpose

+++ Gems/Terrain/Code/Source/Components/TerrainPhysicsColliderComponent.cpp
@@ -183,7 +183,7 @@ namespace Terrain
         numRows = aznumeric_cast<int32_t>((maxBounds.GetY() - minBounds.GetY()) / gridResolution.GetY());
     }
 
-    void TerrainPhysicsColliderComponent::GenerateHeightsInBounds(const AZ::Aabb& bounds, AZStd::vector<int16_t>& heights) const
+    void TerrainPhysicsColliderComponent::GenerateHeightsInBounds(const AZ::Aabb& bounds, AZStd::vector<float>& heights) const
     {
         const AZ::Vector2 gridResolution = GetHeightfieldGridSpacing();
 
@@ -215,7 +215,7 @@ namespace Terrain
                     height, &AzFramework::Terrain::TerrainDataRequests::GetHeightFromFloats, x, y,
                     AzFramework::Terrain::TerrainDataRequests::Sampler::DEFAULT, nullptr);
 
-                heights.emplace_back(azlossy_cast<int16_t>((height - worldCenterZ) * m_heightScale));
+                heights.emplace_back(azlossy_cast<float>((height - worldCenterZ) * m_heightScale));
             }
         }
     }
@@ -287,14 +287,14 @@ namespace Terrain
 
     }
 
-    AZStd::vector<int16_t> TerrainPhysicsColliderComponent::GetHeights() const
+    AZStd::vector<float> TerrainPhysicsColliderComponent::GetHeights() const
     {
         AZ::Aabb worldSize = AZ::Aabb::CreateNull();
 
         LmbrCentral::ShapeComponentRequestsBus::EventResult(
             worldSize, GetEntityId(), &LmbrCentral::ShapeComponentRequestsBus::Events::GetEncompassingAabb);
 
-        AZStd::vector<int16_t> heights; 
+        AZStd::vector<float> heights; 
         GenerateHeightsInBounds(worldSize, heights);
 
         return heights;
@@ -318,9 +318,9 @@ namespace Terrain
         return m_heightScale;
     }
 
-    AZStd::vector<int16_t> TerrainPhysicsColliderComponent::UpdateHeights(const AZ::Aabb& dirtyRegion) const
+    AZStd::vector<float> TerrainPhysicsColliderComponent::UpdateHeights(const AZ::Aabb& dirtyRegion) const
     {
-        AZStd::vector<int16_t> heights;
+        AZStd::vector<float> heights;
         GenerateHeightsInBounds(dirtyRegion, heights);
 
         return heights;
